requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-crispy-forms==2.1